Cookie WPTs: Clean up cookies at the end of each test
This attempts to fix an issue where cookies from previous tests stick
around and interfere with cookie-line test assertions in subsequent
tests, when many tests are run in series.

This adds cookie deletion or cleanup to the end of each test in
wpt/cookies/ which was failing to do so. (There are tests in other
directories that also fail to clean up cookies, but those will be
handled separately.)

Change-Id: I26d0f342e401f4bbcf8ba318683861fa5a45faac